+++ src/main/java/net/geforcemods/securitycraft/blocks/PortableRadarBlock.java
@@ -51,6 +51,11 @@ public class PortableRadarBlock extends OwnableBlock {
 	}
 
 	@Override
+	public boolean isFullCube(IBlockState state) {
+		return false;
+	}
+
+	@Override
 	public EnumBlockRenderType getRenderType(IBlockState state) {
 		return EnumBlockRenderType.MODEL;
 	}