+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/BouncingBettyBlock.java
@@ -42,6 +42,11 @@ public class BouncingBettyBlock extends ExplosiveBlock {
 	}
 
 	@Override
+	public boolean isFullCube(IBlockState state) {
+		return false;
+	}
+
+	@Override
 	public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos) {
 		return new AxisAlignedBB(0.200F, 0.000F, 0.200F, 0.800F, 0.200F, 0.800F);
 	}