+++ src/main/java/net/geforcemods/securitycraft/blocks/mines/MineBlock.java
@@ -43,6 +43,11 @@ public class MineBlock extends ExplosiveBlock {
 	}
 
 	@Override
+	public boolean isFullCube(IBlockState state) {
+		return false;
+	}
+
+	@Override
 	public void neighborChanged(IBlockState state, World world, BlockPos pos, Block block, BlockPos fromPos) {
 		if (world.getBlockState(pos.down()).getMaterial() == Material.AIR) {
 			if (world.getBlockState(pos).getValue(DEACTIVATED))