+++ Changelog.md
@@ -56,6 +56,8 @@
 - Fix: The Keypad Trapdoor's signal length option cannot be set to 0 despite the description saying it can
 - Fix: Console error when changing a Passcode-protected Chest's options
 - Removed: "Is always active" option of Keypads and Key Panels. The same behavior can be achieved by setting the "Signal length" option to 0 and activating the block
+- Fix: Shadow around the Bouncing Betty, IMS, Mine, and Portable Radar
+- Fix: The Bouncing Betty, IMS, Mine, and Portable Radar cause suffocation
 
 --------------------------Changelog for v1.9.8 of SecurityCraft--------------------------